Solve #3 using cmake's ExternalProject mechanism.
Download the Steinberg ASIO SDK via cmake, then
copy the header into the correct place. A scripts
directory has been created for cmake to use as
external commands. Here we now do the substitution
for a 64-bit header, eliminating the need for the
original prepare_64bit_asio script.

Because this uses ExternalProject, it touches
on #4 as well, though we still need an override.